Added entity info and restructured info structs
I still have to implement adding each control flow changing instruction to the currentEntity and give the WALAInstance a way to give the information to the Java, or to create the entities using JNI (rather not).
