+++ .idea/modules/-715730190/swift-source.swan.ca.maple.swan.swift.main.iml
@@ -20,79 +20,7 @@
     <orderEntry type="module-library">
       <library>
         <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.core/build/libs/com.ibm.wala.core-1.5.4-SNAPSHOT-sources.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.core/build/libs/com.ibm.wala.core-1.5.4-SNAPSHOT.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT-sources.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT-javadoc.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT-sources.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT-javadoc.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.util/build/libs/com.ibm.wala.util-1.5.4-SNAPSHOT-javadoc.jar!/" />
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.util/build/libs/com.ibm.wala.util-1.5.4-SNAPSHOT.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
@@ -110,7 +38,79 @@
     <orderEntry type="module-library">
       <library>
         <CLASSES>
-          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.util/build/libs/com.ibm.wala.util-1.5.4-SNAPSHOT.jar!/" />
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.util/build/libs/com.ibm.wala.util-1.5.4-SNAPSHOT-javadoc.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT-javadoc.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.cast/build/libs/com.ibm.wala.cast-1.5.4-SNAPSHOT-sources.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT-javadoc.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT-sources.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.shrike/build/libs/com.ibm.wala.shrike-1.5.4-SNAPSHOT.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.core/build/libs/com.ibm.wala.core-1.5.4-SNAPSHOT.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
+    <orderEntry type="module-library">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../../../WALA/com.ibm.wala.core/build/libs/com.ibm.wala.core-1.5.4-SNAPSHOT-sources.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
