+++ ca.maple.swan.translator/include/SwiftCHook.hpp
